Switch SheetChart/SheetImage dimensions from EMU to pixels
SheetChart.Width/Height, SheetImage.Width/Height, and
CellAnchor.ColumnOffset/RowOffset now take a double in pixels instead
of a long in EMU. The XLSX reader/writer converts to and from EMU at
the file boundary, and all internal pixel math in the overlays drops
its /9525 conversions. Matches the plain-pixel convention used by
ClosedXML, Telerik SpreadProcessing, and Syncfusion XlsIO; avoids
requiring users to think in a unit Excel itself hides.

+++ Radzen.Blazor/Documents/Spreadsheet/SheetChart.cs
@@ -60,14 +60,14 @@ public class SheetChart
     public CellAnchor? To { get; set; }
 
     /// <summary>
-    /// Gets or sets the chart width in EMU (OneCellAnchor only). 1 px = 9525 EMU at 96 DPI.
+    /// Gets or sets the chart width in pixels (OneCellAnchor only).
     /// </summary>
-    public long Width { get; set; }
+    public double Width { get; set; }
 
     /// <summary>
-    /// Gets or sets the chart height in EMU (OneCellAnchor only). 1 px = 9525 EMU at 96 DPI.
+    /// Gets or sets the chart height in pixels (OneCellAnchor only).
     /// </summary>
-    public long Height { get; set; }
+    public double Height { get; set; }
 
     /// <summary>
     /// Gets or sets the chart type.
